ParametrizationCommand in separated file
