+++ src/MiniCover/Commands/BaseCommand.cs
@@ -4,33 +4,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using MiniCover.Commands.Options.FileParameterizations;
 
 namespace MiniCover.Commands
 {
-    internal abstract class ParameterizationCommand : BaseCommand
-    {
-        protected MiniCoverParameterization Parametrization;
-        private readonly IMiniCoverOption<MiniCoverParameterization> _parameterizationOption;
-        protected ParameterizationCommand(string commandName, string commandDescription, params IMiniCoverOption[] options) 
-            : base(commandName, commandDescription, options)
-        {
-            _parameterizationOption = new ParameterizationOption(options.OfType<IMiniCoverParameterizationOption>().ToArray());
-        }
-
-        protected override void AddOptions(CommandLineApplication command)
-        {
-            base.AddOptions(command);
-            _parameterizationOption.AddTo(command);
-        }
-
-        protected override void ValidateOptions()
-        {
-            base.ValidateOptions();
-            _parameterizationOption.Validate();
-            Parametrization = _parameterizationOption.GetValue();
-        }
-    }
     internal abstract class BaseCommand
     {
         protected readonly List<IMiniCoverOption> MiniCoverOptions;
